Add another level of the shader cache
Something like this used to be there in 5.15, and for some time in dev
too, but then it went missing. Reintroduce something similar,
in a safer manner than before, taking the feature set into account
as well. As custom materials already have a proper solution, share some
code with that.

Task-number: QTBUG-86798
Change-Id: Id19fbdbc61aa00521c58a5642cfd00e7b0d265a6
Reviewed-by: Andy Nichols <andy.nichols@qt.io>