Use a simple char array for uniform names
QByteArray becomes evil when the majority of cases pass in a const
char *.

Also reduce the data buffer size, for non-arrays the data is never
larger than 64 bytes.

When it comes to lookups in maps with a QByteArray as the key, use
a wrapping QByteArray which is cheaper to construct (even with an
strlen) than a proper, deep copy based QByteArray. There is of course
an infrequent path when we need to do a proper copy (when inserting
into the map), but the common case should not need any mallocs and
copies.

Task-number: QTBUG-86798
Change-Id: I2dd17d645ec9ac89380a5cf48b762fefafc257cc
Reviewed-by: Andy Nichols <andy.nichols@qt.io>