Fix dangling pointer when removing texture
The problem was that the texture was deleted before the layer causing
the lambda method to set freed variables to nullptr, causing memory
corruption and a crash on exit. Now the destructing lambda is split into
two, one for the manager and one for the texture which makes sure they
are available when destructing.

Fixes: QTBUG-86745
Change-Id: I90162e7930d5f26452930978c429bb9f73e0f679
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>