Skip vertex input gathering for shaders that do not need it
The built-in shaders (skybox, ssao, shadow blurs, etc.) and effect
shaders do not go through bakeVertexInputLocations and do not use
QSSGRhiInputAssemblerState. (because these are typically used with
the quad renderer and that creates its own QRhi input layout, without
using any further abstractions from QSSGRhiContext)

This also means they may want to use arbitrary names for their vertex
inputs in the shader code since it does not really matter for the rest
of the pipeline, they are not connected to mesh attributes.

So have a flag that tells the QSSGRhiShaderStages to skip mapping vertex
input names to some enum.

This also allows removing the special casing of "attr_uv" (which is used
by effects) since now effects can also just set the flag.

Change-Id: I486382cbf3ee3c075699ce1612fd2f7fc2344596
Reviewed-by: Andy Nichols <andy.nichols@qt.io>