Further reduce the chance of having to look up a pipeline
Instead of comparing the srb use the results from the check we
do just before. (the pipeline's srb may have nothing to do with
the dcd's srb, as long as they are layout compatible)

Gain a frame in 1000 x 10k tests. (which are in reality 5000 x 11k
tests but why worry about facts)

Change-Id: If4f87eb52a704ff2fec0f71621ebd6f5ded92c6d
Reviewed-by: Andy Nichols <andy.nichols@qt.io>