Untangle color and specular factor bugs in Principled vs. CustomMaterial
Colors from a 'color' property should be linearized like we do
for colors from a Default/PrincipledMaterial.

Add a note to the CustomMaterial docs about this. (passing in a
color from a _color_ property is linearized, but when writing to
BASE_COLOR, or when basing it on a non-color type property (e.g. just
a vector4d), it is up to the shader)

maybeAddMaterialFresnel contains broken shader code for custom materials,
but that function is never hit for a Principled or Custom material. Make
it clear that it is for DefaultMaterial only.

Finally, take SPECULAR_AMOUNT into account for custom materials when
there is no specular light processor function provided. Passing
qt_material_properties.x to the GGX function is wrong since it should be
qt_specular_factor which is either qt_material_properties.x or
SPECULAR_AMOUNT.

Add a Lancelot case that demonstrates the equivalence of Principled and
Custom materials: first metalness 0 with a non-zero specular amount.
Then something metallic with specular amount set to 0. The left and
right spheres in both rows must be identical, even though one is
Principled while the other is Custom. Finally the bottom sphere should
be identical to the second row, the difference being the linearization
of the color being done in the shader due to not using a QML color property.

Change-Id: If333b024f8c35ff16471ecb80e185d24a1d25ba9
Reviewed-by: Andy Nichols <andy.nichols@qt.io>