Make Custom-Principled identical, default to 0 in metalness everywhere
CustomMaterial {} and PrincipledMaterial {} must be identical.

Similarly, CustomMaterial with a void MAIN() { } fragment shader must be
identical to PrincipledMaterial { }.

This changes the default value of PrincipledMaterial.metalness to 0 as
this makes more sense than 1. It also matches CustomMaterial now.

The DefaultMaterial's specularRoughness defaults to 0 now, instead of
50 (the range is 0..1 to begin with).

SPECULAR_AMOUNT in a custom fragment shader defaults to 0.5 now instead
of 0.

The results of the customsimple Lancelot case change in accordance with
having defaults matching Principled for a custom material without a
fragment shader.

Change-Id: Iead08ad78c0d80d07274b78d16b22896f64c7a18
Reviewed-by: Andy Nichols <andy.nichols@qt.io>