Fix broken light scoping with point and spot lights
This is the reason for the custommaterial example's light probe scene
being broken. The Lancelot cases in the parent patch, esp. the _scope2
one that includes a point light, demonstrates the problem as well.

Skipping writing something to the lightData for point/spot lights is not
a good idea. Presumably having a garbage position or attenuation is
what's causing the problems here.

Change-Id: Ib3fe61aacf54084719335e2ed932cc7b8d300587
Reviewed-by: Andy Nichols <andy.nichols@qt.io>