Make binding lists lighter
A lot lighter in fact. Looking at the generated assembly proves
that we successfully avoid copying sizeof(QRhiShaderResourceBinding::Data)
bytes on generating every element in the bindings list. (with MSVC at least)

The code becomes nicer in fact since we now employ a proper
QSSGRhiShaderResourceBindingList container that offers straightforward
addXx() functions which are now used everywhere instead of filling up
QVLAs manually all over the place. By precalculating a hash value we can
also optimize somewhat the binding list comparisons.

The gains in practice are small, but visible, depending on the scene. It's
somewhere around 1-1.5 FPS in 1000x10k tests on this system.

We are intentionally not using the QRhi static factory helpers anymore
because (1) their non-inline nature is problematic, depending on the compiler,
(2) we want to get a valid QRhiShaderResourceBinding in place, not via a
factory returning a whole new instance. To be seen if the findings here
can be brought back to the QRhi API too at some point, but for now the
goal is to get the hot code path right.

Change-Id: Ie01d60191c458731fe71f85b8fbcc888c06eea0e
Reviewed-by: Andy Nichols <andy.nichols@qt.io>