Rename Hejl-Richard tonemapping to Hejl-Dawson
This appears to be a mistake inherited from the GLTF2 sample viewer
where the tonemapping algorithm authored by Jim Hejl and Richard
Burgess-Dawson is called "Hejl-Richard". Other sources call it
Hejl-Dawson, which is more in accordance with how algorithms are
usually named after its authors. Using the same name as these other
sources will make it easier to find information about the algorithm.

Change-Id: Ia16b72eec947acfaae721716ce6f7ab1f6b61579
Reviewed-by: Christian Strømme <christian.stromme@qt.io>