doc: Simplify references to BufferInput/TextureInput in Effect docs
There was some seemingly unnecessary complexity when referring to
BufferInput and TextureInput in the Effect documentation, so this is
an attempt at simplifying it.

First, there was a list of three ways to use BufferInput. Two of the
ways were identical, except that one used the default assignment for
one property. The third was to reference a TextureInput, which the
documentation immediately after explained was redundant, since
result would be exactly the same even if the BufferInput was omitted.

It seems like the documentation was written to explain all the possible
code paths internally rather than explain how the user should approach
these types. This is an attempt at turning it around and limit the
information to what the user needs to know.

Change-Id: I8573beec9f95e5745e180cc0fd22418020502de7
Reviewed-by: Andy Nichols <andy.nichols@qt.io>