Finish modernizing import statements in examples
This amends 9cdc42156b338f5d633c1261c7f1a217b2fea019, which
removed versioned imports as a drive-by in files that
contained the outdated import of QtQuick.Window. The remaining
ones were not touched in that commit.

With recent changes in Qt Quick Controls 2, though, importing
older versions will fail on macOS due to a failure on loading
the macOS style. So for the examples to run successfully there,
we need to remove the versions everywhere.

Fixes: QTBUG-87653
Change-Id: Ida1ef1f2228441f1c9d1a27a0a3a4fe3afd0d734
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>