doc: Small update to Loader3D
Most of the documentation here is copy-pasted from Loader, so in the
interest of keeping them consistent, they should be left as they are.

This change simply makes the difference between Loader and Loader3D
more explicit, by explaining that it is the superclass of the
components that decides which one to use.

Change-Id: I7d0b115f1fd3f59826ace7c58bce9efd9ca8a23e
Reviewed-by: Christian Strømme <christian.stromme@qt.io>