Do not crash with non-HDR images for lightProbe
This attempts to resolve the longstanding controversy around the
inFormat argument. (see todo note)

For QImage-based formats and compressed formats the inFormat argument
is now not present at all. This should be ok since inFormat is almost
always UnknownFormat, except when the image is used as a light probe.
But if it's a light probe, then the inFormat hint only makes sense if
it's a .hdr file. It is not applicable to a QImage or compressed
format. (because there the input file defines the format, not our
wishes)

By dropping the inFormat argument in the QImage-based image loading
path, we avoid crashing when a .png or other QImage-based format is
specified in the lightProbe property. Previously this led to attempting
to create a RGBA16F QRhiTexture with RGB(A)8 image data from a QImage,
which led to various unexpected behavior.

Task-number: QTBUG-88005
Change-Id: I4d5a873d8246b65fe8074e135a66a2f89b3f0ac8
Reviewed-by: Andy Nichols <andy.nichols@qt.io>