Remove QtQuick3D.Helpers.PointerPlane
This API doesn't really match anything else in the API and is not
currently in use by any example (nor in Qt Design Studio), so in order
to rethink it, we remove it for now.

It seems to originally have been written for the studio example,
which was later removed because "it wasn't really showing anything",
and the use case seems to only be to calculate intersections from
viewport rays with a plane in the scene, which should probably be
a method on the camera and not a node in the scene graph, to match
the other APIs that do similar things and avoid clutter.

Change-Id: If0c4c692d2e807029af970a209896d43c9dd9521
Reviewed-by: Andy Nichols <andy.nichols@qt.io>