Sample base/diffuse color map once only
As powerful and high performing Qt is, it still cannot afford to sample
these textures twice for no reason.

This has to include the alpha cutoff fixes as well since the code in
question has to be moved anyway.

- The test with alphaMode==Mask is now performed early on, regardless of
  where the base color alpha is coming from (the color or the texture map).

- In addition, the test matches the glTF spec now: when the test fails,
  the alpha is forced to 1.

- For diffuse maps on DefaultMaterial the alpha is no longer multiplied in
  twice.

All this changes the results in certain Lancelot tests: the two
alphaMode related tests change here and there as expected, while the
'compressed' test changes due to the DefaultMaterial diffuse map alpha
fix.

Done-with: Inho Lee <inho.lee@qt.io>
Fixes: QTBUG-88103
Change-Id: I5f7b97874bd2f996bdb4acc913dde1cbfb543a38
Reviewed-by: Inho Lee <inho.lee@qt.io>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>