Remove LIMITBONEWEIGHTS_PROCESS when building assimp
This post processing step limits bone weights size for a vertex into
maximum value(0x4) and remove bones having empty weights.
The problem is that when assimp removes a bone in a mesh, we cannot
figure out the corresponding node is a bone node.(It's a limitation of
assimp.)
We are already checking empty weights and limit the maximum number.

Change-Id: I6057a251725746b3a7dea3b20e3d8ed200da34f4
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>