Fix normal map without uv coords shader crash
In the shader, the normal map uses qt_varTexCoord to retrieve its
'normal' value but if no uv coords are specified then qt_varTexCoord
does not exist and the shader will fail to compile. We work around this
by setting qt_varTexCoord to zero if it does not exist as an input to
the vertex shader.

Fixes: QTBUG-86374
Change-Id: I7a3e64295832750281dd227fbd25371b91c31a9d
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>