Remove warning about ref count in QQuick3DObject
The warning is ominous, but the fact is that we don't actually care
at this point, it's too late anyways. If we add proper tracking and
tear down, where it's actually expected that ref returns to 0, then
the warning should be reintroduced. The warning was added in
c39067f20be92

Change-Id: Ibef2c60269e4c9e00c65743cf84af48af55d2f7c
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>