Expose Effect type as part of QtQuick3D import
This was part of the QtQuick3D.Effects import before but the Effects
import is the premade effect library, and is not necessary to create new
Effects. All of the plumbing and definitions of Effects are part of the
QtQuick3D module, so it should be exposed there.

This makes QtQuick3D.Effect a pure QML import.

Fixes: QTBUG-88133
Change-Id: I6b3518d41e2e71a7634943397fdb836e147824e7
Reviewed-by: Christian Strømme <christian.stromme@qt.io>