Fix issue with import scenes
Avoid breaking the list links of imported scenes.
We would break the node list of imported scenes when we blindly added
them to the child list of the layer. Touching the list pointers of
another list is a no-go which will can cause the list to go into a
weird state, e.g., disconnected or looping.

Fixes: QTBUG-88487
Change-Id: Icc2b4f80e1ad1c96fa2dd4e1da78c89e9e431652
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>