Support input events in 3D
This adds support for forwarding input events in Qt Quick to
2D items in 3D scenes. There are two paths possible where you would
want to forward events from 2D -> 3D -> 2D.
- 2D Items embedded in 3D scene
    These items are rendered directly into the scene with the correct
    3D projection.  QQuick3DItem2D ctor calls
    ensureSubsceneDeliveryAgent()
    on the new root item so that it becomes the root of a subscene.
- 2D Items used as texture sources
    These items are rendered offscreen and used as texture sources.
    This adds the extra complexity that events may come from any object
    in the scene that uses this special texture in its material.
    ensureSubsceneDeliveryAgent() is called on the source item.
To faciliate this extra automatic picking logic is needed.  This picking
code is responsible for checking if any items in the 3D scene contain
the two above cases.  In addition when ray casting, we care about all
intersections and go through each intersection in order of distance to
the camera, stopping only when the event is accepted.

After a 2D Item grabs an eventpoint, QQuickDeliveryAgent receives the
next event and delivers it directly to the grabber, which would bypass
the picking logic.  But we call QQuickDeliveryAgent::setSceneTransform()
ahead of time to provide a functor (ViewportTransformHelper) that knows
how to map the 2D viewport position to the 2D subscene position, by
using QQuick3DSceneRenderer::syncPickOne() to repeat the UV mapping.
This is faster than whole-scene picking, because the relevant 3D object
is already known.

However, whole-scene picking occurs during delivery of each hover event.
Hover events can be disabled to avoid that overhead.

Task-number: QTBUG-84870
Pick-to: 6.1
Done-with: Shawn Rutledge <shawn.rutledge@qt.io>
Change-Id: Ia9760f5f7bded4b185413222ec428b60e657cc1c
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>