Reduce flicker for non-layer Item trees in Texture.sourceItem
This tends to work very well when referencing a standalone, invisible
item tree sourceItem (Item { id: stuff; visible: false; ... }), in my
tests it seems to give immediate response both at startup (in the first
frame) and when changing sourceItems.

With inline item trees (f.ex. sourceItem: Rectangle { ... }) things are
improved as well, but that still has cases where we do not get an immediate
update, at startup in particular. In the manual test this is now
exercised by the cone, which only gets valid 2D content into the texture
on the 3rd frame for some reason. The other Texture(s) used by the other
three models get content in the very 1st frame. This has something to do
with inline Item trees not getting a corresponding QSGNode tree early
enough. (this is likely because we refWindow'ing only in updateSpatialNode,
which then means that first frame is out of question, then in the second
frame we still have nothing when updating the texture in beforeSynchronizing,
then in the Item sync round it gets QSGNodes in updatePaintNode, which is then
finally picked up when rendering the texture for the first time in the 3rd frame;
this should be investigated in a follow up) In some other examples of
using inline sourceItem content, the improvement is however there
(perfect first frame, instead of only getting correct texturing in the
fourth)

Task-number: QTBUG-87077
Change-Id: I3566474f45fc71089f8342b158fb16014d667306
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit c240df59cbb468ac43945dab3dd5337ff318479f)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>