Do a refWindow as early as possible for Texture's sourceItem
With this we tackle the last remaining problems with sourceItem texture
generation getting deferred: now even if an inline item has no window
set, we do a refWindow earlier than Texture::updateSpatialNode (though
we still keep the logic there as a fallback is all else fails), by
connecting to the sceneManager's windowChanged when receiving the
ItemSceneChange. This way when we get to synchronize, and to
Texture::updateSpatialNode, the sourceItem is already associated with a
window, and presumably has QSGNodes created, and thus can get content
rendered via QSGLayer right away.

As usual, there is no corresponding derefWindow. This is a potential
problem, but no different from Texture's current behavior.

Task-number: QTBUG-87077
Pick-to: 6.0
Change-Id: I60e00e129a363d799dd745697b7563665b885dbb
Reviewed-by: Andy Nichols <andy.nichols@qt.io>