Introduce the base for autotests with on-screen rendering
Introduce visual autotests that bring up the full stack, meaning
at minimum an on-screen QQuickView with a View3D in it.

Unlike Qt 5.15, where Quick3D's OpenGL requirements were somewhat
problematic for CI use, in Qt 6 we can run QRhi-based rendering in
the CI virtual machiness, even though in practice this means running
with a software rasterizer: D3D11 with WARP on Windows, Mesa llvmpipe
on Linux, and the "Apple software renderer" with OpenGL on macOS.
Vulkan and Metal are not exercised in any form in the CI at the time of
writing. (macOS would default to Metal normally but falls back to OpenGL
in the VMs)

Such tests are not a replacement or alternative to the Lancelot-based
tests. It is expected that these autotests will limit themselves to
simple scenes, and avoid advanced graphics features (although most,
if not all, things should be available in CI virtual machines as well,
since OpenGL 3.x should be mostly sufficient for Quick3D atm)

On the other hand, these tests can exercise dynamic changes in the
2D and 3D scene, or on the window, and can open and test multiple windows,
exercise multiple viewports, simulate input events, etc. That is where
the real value is here.

Follow the patterns from Qt Quick autotests such as scenegraph or
qquickwindow: the typical test will create one or more QQuickViews
using the shared helper code and .qml files in the 'data' folder.

Once the window is exposed, it can change state, and grab into a
QImage. Proper visual verification is not in scope: we should be fine
with a few simple pixel checks because our goal is not to verify the
scene is rendered 100% correct visually, but to see if the scene can be
rendered at all, and that dynamic changes lead to changes in the scene,
the underlying objects, and in the visual output. There is always a
chance of unstability when checking pixel colors, so leaving some room
for difference is recommended. This system is and will not perfect,
but significantly better than nothing.

We start with two tests:
- the smoketest is a simple bringup test that has
  some additional debug prints enabled, to help troubleshooting when
  looking at CI logs for example.
- simplescene currently just has one simple test rendering a cube, and
  serves as the base example for further, more useful test cases in the
  future.

The helper code, based on Qt Quick's, is extended with some common
functions for creating a QQuickView and grabbing its content into a
QImage. For troubleshooting (or just to see what the actual rendering
result is) set the environment variable QT_QUICK3D_TEST_DEBUG=1. This
enables saving each grabbed QImage into a png file. For convenience the
base initTestCase performs the typical checks: running with offscreen
or minimal platform plugin, or having Quick use a non-RHI rendering path
will skip all tests. Hopefully this will be sufficient for not having to
replicate such obscure checks in every test.

Pick-to: 6.0
Change-Id: I60fe918c33bada5ceddc3a84d0b71cb0971addf5
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>