Add a rendercontrol-based autotest
This ensures that basic functionality is available in combination with
an invisible QQuickWindow + QQuickRenderControl, which is essential
for Design Studio, as well as for other use cases. (however, it has to
be kept in mind that Metal and Vulkan are not currently tested in the
CI VMs, which is unfortunate, but not much we can do about it now; the
test is prepared to function with all QRhi backends regardless)

This also gives an opportunity to exercise Texture.sourceItem in a
controlled manner: we can make sure the 2D content rendering is not
unexpectedly deferred since with rendercontrol we are in control of when
a frame is generated, whereas with an on-screen QWindow there is always a
risk of having more updates than expected due to varying windowing system
behaviors.

Pick-to: 6.0
Change-Id: I3d58eb4d159cd24c8dff5a16e18b02f193d9eeb5
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>