Fix transparency handling with QSGTexture
Setting the QSSGRenderImage::m_textureData's m_textureFlags is futile
unless the loadRenderImage overload preserves it. For QSGTextures this
is not the case. Rather, change the loadRenderImage variant to set the
transparency flag based on QSGTexture::hasAlphaChannel(). (which is
in practice always true for a QSGLayer, while it depends on
QImage::hasAlphaChannel() for a QImage-based QSGPlainTexture, i.e.
when the source is an Image item)

Task-number: QTBUG-88775
Change-Id: Ibca549d81eeefd29b230796007d7c090e9b53a87
Reviewed-by: Andy Nichols <andy.nichols@qt.io>