Clean up index vs. light index confusion
Dynamically toggling the visibility of shadow casting lights can,
depending on what order lights are specified in the QML object tree,
lead to assertions (or in release builds, crashes).

The key problem here is addShadowMapEntry treating the 'index' as an
index into the vector, whereas it is meant to be a lookup key (light
index), independent of the underlying container. In addition to
correcting this, clean up the terminology a bit.

Fixes: QTBUG-89062
Change-Id: Ieb0f1b0bcbe02d81dbf482484f059a0fb7f13b81
Reviewed-by: Christian Strømme <christian.stromme@qt.io>
(cherry picked from commit 50afefe8d1e3d79d4ef973efd171bf3133ff9eed)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>