Prefix the shader key for effects and custom materials
This is similar to what we do for materials. Unlike the built-in
materials, the effect and the custom material does not have a
unique key based on the properties of the pipeline, so the file
path was used instead. Using the file path has its own problems as the
caching of the pipeline code is now tied to a specific file path, but
also for the offline shader generator, which might use different path
for the input files (shaders, texture etc.) as the resource file is
not available for the shadergen tool.

Pick-to: 6.0
Change-Id: I7c9622b5712e0c3cdb47ddc72cb3e08c4f7118ac
Reviewed-by: Andy Nichols <andy.nichols@qt.io>