Show a warning once upon too many shadow casting lights
Just follow what we do for the other limit on lights.

There are some shady details (what exactly counts as
a shadow casting lights, e.g. visible is not taken into
account here, but that's consistent with the max light
limit in this regard (15 is not the limit for active
(visible) lights, rather, it is the limit for all lights
due to intricate details of how the shader generator works
wrt lights)

Pick-to: 6.0
Change-Id: I531490c9dc1353b62c55831347dc6d5de163cbf6
Reviewed-by: Andy Nichols <andy.nichols@qt.io>