Fix ref counting in prepare functions
Some of the prepare functions were taking the pointer for the context
interface and increasing the ref count twice, or more, for no good
reason. The primary reason for this is that the return type from the
renderer changed at some point.

Change-Id: I73cec851f469fb77de005e5469434f8d88cffea8
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
(cherry picked from commit 3084ef3311fccfef5821fd353de98ddb55a8c298)
Reviewed-by: Andy Nichols <andy.nichols@qt.io>