Improve blendMode docs
The default SourceOver _blend_ mode is quite confusing given that
it does not imply alpha blending on its own. Clarify this. For
PrincipledMaterial mention the interaction with alphaMode as well.

Leave the other two mode docs as-is for now, text is untouched but is
reformatted.

Change-Id: I1c9d7f4eecd908ddf02ec7e5ab95c23c09c789d2
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit 31074aa74d5bfdf0645f801faa0ff4b72bc9e373)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>