Switch Mask alphaMode to use a true discard
Which in turn allows us to treat such objects as opaque by default.
(because the cutoff test either succeeds in which case the alpha is
forced to be 1, or fails in which case the fragment is discarded -
previously the alpha was changed to 0 which requires alpha blending
to produce appropriate results, but that's against the glTF spec in
practice, and also makes it not possible to cast shadows)

Now, there may be performance implications (discard destroys early Z
etc. etc.), but in practice there isn't much else to do.

Fixes: QTBUG-82331
Pick-to: 6.0
Change-Id: Id462976e35586774f7f1eb35a96b1cb50cc0d083
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
Reviewed-by: Inho Lee <inho.lee@qt.io>