AssimpImporter: Sync attenuation constants with quick3d backend
Quick3D's attenuation constants are applied with multiplied 0.01 for
the linear one, and 0.0001 for the quadratic one.
And for constantFade, glTF's recommendation is using just quadraticFade,
so it seems Assimp to make it as 0. In that case, attenuation goes
infinite when distance is 0. So we will use 1.0 constantFade in that
case.

Change-Id: I7c19edc2d43cb05096e68f2984745e64dfe1d3ee
Reviewed-by: Andy Nichols <andy.nichols@qt.io>