AssimpImporter: Add the value of the brightness for Lights
Assimp has no property related to brightness or intensity and just
provides intensity-multiplied colors. If a light has somehow large
intensity value, the light colors and ambient colors falls into "white".
This patch will extract the maximum color component and use it as the
'brightness' to prevent this wrong operation.

Change-Id: I487e59171832689868160f532db70ee78689bdc5
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>