Shadergen: Clean-up the type info
This is an attempt to make the type system a bit more sane and easier to
use. Instead of having additional type values defined, we now use the
QMetaType system. There are some downsides with not having our own type
identifiers, as it means we don't have compile time type ids for the
non-builtin types (for now).

Change-Id: I675dbd1c63c500fb34ea2ff7ae2a33aa301d722d
Reviewed-by: Andy Nichols <andy.nichols@qt.io>