Fix warning about unused variable after fix to remove unused variable
This amends ba153f2014c66 where the argument to prepareForRender()
ended up unused. This is a quick fix done by adding a Q_UNUSED()
around the now unused argument. A more correct solution will take
more effort as the function is virtual and in one of its incarnations
the argument is actually used.

Change-Id: I756bacf3409735af0d488ee3ebc50ecde3761578
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>