Remove a bogus and unused stride calculation with custom geometry
This code was never hit since we do require calling
QQuick3DGeometry::setStride(). Which is good because it seems what it
calculates is not the vertex attribute stride, but something else.

With an arbitrary vertex buffer as the input, where the caller is
responsible to decide how the vertex data is laid out (interleaved
or not, random empty space between data for the different
attributes, etc.), it is not possible to calculate the vertex
attribute stride without knowing the layout. Just remove this,
and add a warning in QQuick3DGeometry if the stride is not set,
just in case.

Change-Id: Ib46740943bbdbe0ce808e32c72d1fbdae42be1bf
Reviewed-by: Andy Nichols <andy.nichols@qt.io>