Update commercial license headers
Updated header.COMM to the files in tqtc-qtquick3d. Examples, tests
or documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals with the
previous license header. Reason for this is that some autotests
use hard coded line numbers and a change in the line count
causes failures in tests.

Task-number: QTQAINFRA-4224
Change-Id: Iee29706b3f4288ef75570213fa281ed2c28fef0f
Reviewed-by: Antti Kokko <antti.kokko@qt.io>