Purge existing, inherited, unused-in-practice lightmap stuff
This is a public API break in a way because we will no longer have the
three lightmap texture properties, but said properties should not be in
use in practice by any real world applications.

Said properties are documented as
"This feature is still in development so there is currently no way to bake lights."
so removing them should cause no problems in practice.

In the uip converter we still parse the properties from the .uip file
but do not generate the corresponding code in the .qml file.

Change-Id: I4c6e53298990c0a7ff5b0b20d8b1aa18fffd5681
Reviewed-by: Andy Nichols <andy.nichols@qt.io>