Add new Mesh writer and builder
This will bump the mesh version to 4. The loader supports version 3 and
4 without any special branching because the only difference in version 4
is that the "offset" values are zeroes. We still need a new version just
to prevent a previous Qt version to try to load the .mesh files written
by a newer version.

Calculating the offset values (and so writing out valid V3 files) is not
reasonable: the original mesh code, defying all practices for readable,
maintanable, and robust code, makes it difficult to reproduce the same
data in a sane serializer, meaning calculating offsets would need baking
in legacy logic which is not affordable at this point.

MeshBuilder is replaced by a single static from function, similarly to
what we already have for custom geometry. Due to using plain and
simple C(++) structs within the new Mesh object, much of the old
MeshBuilder's logic is just not needed. The assimp importer has been
migrated to the new types.

The default index type is back to uint16. Not clear why the default was
changed to uint32 at some point, even when the vertex count is well below
64K. Likely due to a bug in the code: the / 3 divide is now removed so
it should be good.

The OldMesh namespace is now gone.

qssgmeshutilities.cpp/_p.h are now qssgmesh.cpp and _p.h because that's
what it really is: the implementation of Mesh.

The NewMesh namespace is now QSSGMesh.

The patch also improves mesh reading, by avoiding seek() calls as much as
possible, instead just reading the padding bytes. The alignment padding
is also a legacy, completely unnecessary part of the format, but we keep
that in order not have to branch in the reader based on the mesh
version.

Change-Id: Ic83cd69360d0d66801e5aae86bbfd9dfe234a321
Reviewed-by: Andy Nichols <andy.nichols@qt.io>