Instancing: Let default transform origin be intuitive
Previously, the position of a model would be applied locally, so that a
rotation of the instance transform would cause the instance to orbit
the position from the instance table. This is not what you would
generally expect, and not what is needed for particles. On the other
hand, applying the model's rotation and scale globally would cause the
entire set of instances to be transformed in an unexpected way.
Therefore, by default we do 'magic' and remove the translation part from
the local instance transform and apply it globally.

To avoid the magic, set the instanceRoot property explicitly. As part of
this change, the logic is fixed so it is now possible to set a model's
instanceRoot to itself.

Change-Id: Ic41bf1bea57222f3fbd6a1d732489ae400a8f0e8
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit 13ef6c88aca7f0217103718517970efb9355a291)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>