More scene manager fixes
Simplify the window and render context binding code (tested against
QTBUG-87678).
Move the cleanupResource() function out of the scene renderer and into
the viewport class and remove the need for the scene renderer to hold
onto handles to the scene, import scene and the window, all which
the viewport already has an intimate relationship with.

Fixes: QTBUG-90514
Pick-to: 6.1 6.0
Change-Id: Ie470cb9a65149a58f0bedcd2251801f8a30460e0
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>