CodeChecker: QQuick3DSceneRenderer bail when there is no command buffer
I'm not sure if this is the correct thing to do here, because if we
can't find get a valid command buffer, then no work can be done.
Previously we print a warning and crash. At this point we might crash
somewhere else. Another option is that we could return nullptr for the
texture, but we do have a texture to return at this point, we just
didn't get a chance to render to it.

Pick-to: 6.1
Change-Id: I0230b1e9e7f2eb11cf6a42dc5c70c0e052243b46
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>