CodeChecker: GridGeometry be explicit when using integer division
In this case we are using the behavior of integer division explicitly
and constructing QVector3D with int. Because it accepts floats there is
an implicit conversion, and mixing that with integer division triggers
CodeChecker as a possible error. By explicitly casting to floats we make
it clear this is intentional.

Pick-to: 6.1
Change-Id: I32291e1e4b267edc879e5161410396c294754669
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>