Show something on QSG_RENDERER_DEBUG=render
Just as the Qt Quick batch renderer prints some stuff when this is set,
print something from the 3D renderer as well. Due to pretending having
a unified 2D-3D renderer to the outside world, attempts to derive data
about the amount of geometry drawn can now end up in misleading and
wrong data when View3D is involved. Remedy this by simply plugging
into the same env.var. and doing some qDebugs when enabled. The result
is a nicely interleaved output from both the 2D and 3D renderers.

This way one gets a better impression on the draw calls and the number
of indices/vertices involved in the entire 2D+3D environment. We can also
show something about the number of render passes and the size of the render
targets, thus encouraging people not to use shadows for example.

Do it the right way and make sure nothing gets collected when the
env.var. is not set (we do not want to spend time on counting either).

Change-Id: I71b82204e63507bdd9b8a9a9668c8ffd31caf7ce
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit 234e70d98712bc2010330f9cc28eac44b00d0e66)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>