CodeChecker: QQuick3DParticleSystem don't use fabs on a char
These values may have been QVector3D's previously, but now they are
Vector3b's, which are apparently vec3<char>. char is signed 8 bit ints,
so use std::abs instead of std::fabs.

Also not going to change it in this commit, but we should have used
qint8 or int8_t insteaed of char, even if they mean the same thing. The
compiler doesn't care, but the next person reading it will.

Pick-to: 6.1
Change-Id: Idf569e700de3aead4b0f5c373a717e9eac7636aa
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>