CodeChecker: QSSGRenderShadowMap prevent deref of null pointer
Just looking at the code it *shouldn't* be possible to hit this case
currently, but CodeChecker thinks it *could* and if we did it would
crash. This else condition only allies to the CUBE case now, so it
doesn't hurt much to add a conditional for this as well.

Change-Id: I250dc16f56f865c1111631610ad0c80bee2c38c2
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
(cherry picked from commit c98bcb6b3d7bedf90c95deaaf5fec75a648cf5ce)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>