Add lightmap UV generator
When --generateLightmapUV is not passed to balsam, there is no change
in the results whatsoever, apart from a performance increase in the
importing process due to avoiding redundant QByteArray creations
(mallocs and deep copies).

When --generateLightmapUV is specified, attr_uv1 will be the lightmap UV
channel. If there is UV1 in the source mesh and lightmap UV is enabled,
the mesh's UV1 is ignored. To be decided later on if UV1 is the right
choice or we want to introduce UV2.

The lightmap size hint is not yet stored in the QSSGMesh, that is to be
implemented separately.

Task-number: QTBUG-89863
Change-Id: If2b0d80218b3a390619b2381567f85424125951f
Reviewed-by: Andy Nichols <andy.nichols@qt.io>