Make Texture.indexUV: 1 safe to use and add a Lancelot test
At the moment setting indexUV to 1 on a Texture that is then referenced
by a material may or may not work: if there is another material that
matches ours (when it comes to material keys), the indexUV will be
ignored because nothing ensures our material, where only the indexUV is
different, gets its own shaders generated.

The indexUV must be part of the key, generated in prepareImageForRender.
For now just have a bool indicating that the image wants UV1, not UV0.

Fixes: QTBUG-91118
Change-Id: Ib9b4dfa6b5b442c4629847cce12c963077612c39
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit a7df91a12c50725abab9c0d18539d686615bb8b5)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>