Fix to deal with emissive properties
The descriptions about emissive term in Quick3D materials are not
clear. This patch makes them to match with glTF2's concept.

The emissiveColor is removed in both default and principled materials.
The emissiveFactor becomes a vector3d channel multiplier in both materials.
The emissiveMap is a RGB texture to specify the intensity of emissive
factor's each channel.

[ChangeLog][QQuick3D][QQuick3DMaterial/Emissive] This patch changes how
to handle emissive terms in all materials.
Following lancelot tests are influenced by this patch.
custommaterial/customsimpletexture.qml
instancing/customsimple.qml
layers/Light_probe.qml
Previously the emissive term is multimpled by the diffuse term, but now
it is added to the final color.  In customsimpletexture, the final color
will be close to "white" since the emissive value is too big.

Change-Id: Ib03dcaa6fb9eb8a70bff9c3fcf8454408c62411a
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>